+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 **Changed**
 - Move formatters from `Dicey::OutputFormatters` to `Dicey::CLI::Formatters`. No longer `require` them by default.
 - Rename `HashFormatter` to `BaseMapFormatter` and `KeyValueFormatter` to `BaseListFormatter`.
+- CLI can now be enabled with `require "dicey/cli"` instead of `require "dicey/cli/blender"`. CLI calls can be made with `Dicey::CLI.call([argv])`.
 
 [Compare v0.16.2...main](https://github.com/trinistr/dicey/compare/v0.16.2...main)